Moved some classes for clarity
