+++ src/main/scala/devsearch/features/Features.scala
@@ -1,5 +1,7 @@
 package devsearch.features
 
+import devsearch.ast._
+
 abstract class Feature(position: CodeFilePosition) extends java.io.Serializable {
   def key: String
 }
@@ -7,3 +9,27 @@ abstract class Feature(position: CodeFilePosition) extends java.io.Serializable 
 trait FeatureExtractor extends java.io.Serializable {
   def extract(data: CodeFileData): Traversable[Feature]
 }
+
+object FeatureExtractor extends java.io.Serializable {
+
+  lazy val extractors = List(
+    ImportFeatures,
+    InheritanceFeatures,
+    ValDefFeatures,
+    FunDefFeatures,
+    TypeFeatures,
+    StructuralFeatures
+  )
+
+  def extract(data: CodeFileData) = extractors.flatMap(_.extract(data))
+}
+
+case class CodeFileLocation(user: String, repoName: String, fileName: String) extends java.io.Serializable {
+  def at(pos: Position) = CodeFilePosition(this, pos.line)
+  def at(line: Int) = CodeFilePosition(this, line)
+}
+
+case class CodeFilePosition(location: CodeFileLocation, line: Int) extends java.io.Serializable
+
+case class CodeFileData(location: CodeFileLocation, ast: AST) extends java.io.Serializable
+
